pyproject.toml
[project]
name = "explainableacd"
version = "0.1.0"
description = "Explainable Automatic Claim Detection"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.75.0",
    "baml-py>=0.70.0",
    "groq>=0.9.0",
    "contractions>=0.1.73",
    "dspy>=2.6.0",
    "emoji>=2.15.0",
    "fasttext-wheel>=0.9.2",
    "ftfy>=6.3.1",
    "joblib>=1.5.2",
    "marimo>=0.18.4",
    "mlflow>=3.7.0",
    "numpy<2",
    "openai>=1.0.0",
    "plotly>=6.5.0",
    "polars>=1.36.1",
    "pydantic>=2.0.0",
    "pyarrow>=22.0.0",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.8.0",
    "sentence-transformers>=5.2.0",
    "spacy>=3.8.11",
    "together>=1.0.0",
    "tqdm>=4.66.0",
    "wordsegment>=1.3.1",
    "xgboost>=2.0.0",
    "lightgbm>=4.0.0",
    "catboost>=1.2.0",
    "optuna>=3.5.0",
    "langdetect>=1.0.9",
    "umap-learn>=0.5.9.post2",
    "einops>=0.8.1",
    "json-repair>=0.54.3",
    "imbalanced-learn>=0.14.1",
    "accelerate>=1.12.0",
    "transformers>=4.57.3",
    "datasets>=4.4.2",
    "sentencepiece>=0.2.1",
    "nltk>=3.9.2",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "ty>=0.0.1a1",
]

[tool.ruff]
line-length = 120
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"notebooks/*.py" = ["N807", "N803"]
"experiments/scripts/*.py" = ["E402"]

[tool.ty.environment]
python-version = "3.13"

[tool.ty.rules]
# Suppress warnings for third-party libs without type stubs
possibly-missing-attribute = "ignore"
